refactor(core): modularize MessageList class into specialized components (#11658)
