+++ packages/core/src/agent/index.ts
@@ -1,5 +1,5 @@
 export { TripWire } from './trip-wire';
-export { MessageList, convertMessages } from './message-list';
+export { MessageList, convertMessages, aiV5ModelMessageToV2PromptMessage, TypeDetector } from './message-list';
 export type { OutputFormat } from './message-list';
 export * from './types';
 export * from './agent';
