+++ packages/core/src/agent/message-list/adapters/index.ts
@@ -1,0 +1,3 @@
+export { AIV4Adapter, type AIV4AdapterContext } from './AIV4Adapter';
+export { AIV5Adapter, type AIV5AdapterContext } from './AIV5Adapter';
+export type { AdapterContext, UIMessageWithMetadata } from './types';
