+++ packages/core/src/agent/message-list/cache/index.ts
@@ -1,0 +1,2 @@
+export { CacheKeyGenerator } from './CacheKeyGenerator';
+export type { MastraMessageContentV2, MastraMessagePart, UIMessageV4Part } from './types';
