+++ packages/core/src/agent/message-list/detection/index.ts
@@ -1,0 +1,1 @@
+export { TypeDetector, type MessageInput } from './TypeDetector';
