+++ packages/core/src/agent/message-list/merge/index.ts
@@ -1,0 +1,1 @@
+export { MessageMerger } from './MessageMerger';
