+++ packages/core/src/agent/message-list/state/index.ts
@@ -1,0 +1,12 @@
+export { MessageStateManager, type MessageSource, type SerializedMessageListState } from './MessageStateManager';
+export type {
+  MastraDBMessage,
+  MastraMessageV1,
+  MastraMessageContentV2,
+  MastraMessagePart,
+  UIMessageV4Part,
+  UIMessageWithMetadata,
+  MemoryInfo,
+} from './types';
+export { serializeMessage, deserializeMessage, serializeMessages, deserializeMessages } from './serialization';
+export type { SerializedMessage } from './serialization';
