+++ packages/core/src/stream/base/output.ts
@@ -1,7 +1,8 @@
 import { EventEmitter } from 'node:events';
 import { ReadableStream, TransformStream } from 'node:stream/web';
 import { TripWire } from '../../agent';
-import { MessageList } from '../../agent/message-list';
+import { coreContentToString } from '../../agent/message-list';
+import type { MessageList } from '../../agent/message-list';
 import { MastraBase } from '../../base';
 import { ErrorCategory, ErrorDomain, MastraError } from '../../error';
 import { getErrorFromUnknown } from '../../error/utils.js';
@@ -659,9 +660,7 @@ export class MastraModelOutput<OUTPUT extends OutputSchema = undefined> extends 
                   // Get text from the latest response message (the last assistant message)
                   const responseMessages = self.messageList.get.response.aiV4.core();
                   const lastResponseMessage = responseMessages[responseMessages.length - 1];
-                  const outputText = lastResponseMessage
-                    ? MessageList.coreContentToString(lastResponseMessage.content)
-                    : '';
+                  const outputText = lastResponseMessage ? coreContentToString(lastResponseMessage.content) : '';
 
                   // Only update the last step's text if output processors actually modified it
                   // This preserves text from retry scenarios where step.text is already correct
