+++ stores/_test-utils/src/domains/memory/messages-list.ts
@@ -2,7 +2,7 @@ import { beforeAll, beforeEach, describe, expect, it } from 'vitest';
 import { createSampleMessageV2, createSampleThread } from './data';
 import type { MastraStorage, MemoryStorage } from '@mastra/core/storage';
 import type { MastraDBMessage, StorageThreadType } from '@mastra/core/memory';
-import { MessageList } from '@mastra/core/agent';
+import { MessageList, TypeDetector } from '@mastra/core/agent';
 
 export function createMessagesListTest({ storage }: { storage: MastraStorage }) {
   let memoryStorage: MemoryStorage;
@@ -78,7 +78,7 @@ export function createMessagesListTest({ storage }: { storage: MastraStorage }) 
 
       expect(result.messages).toHaveLength(5);
       expect(result.total).toBe(5);
-      expect(result.messages.every(MessageList.isMastraDBMessage)).toBe(true);
+      expect(result.messages.every(TypeDetector.isMastraDBMessage)).toBe(true);
     });
 
     it('should list messages with pagination', async () => {
@@ -373,7 +373,7 @@ export function createMessagesListTest({ storage }: { storage: MastraStorage }) 
         threadId: thread.id,
       });
 
-      expect(result.messages.every(MessageList.isMastraDBMessage)).toBe(true);
+      expect(result.messages.every(TypeDetector.isMastraDBMessage)).toBe(true);
     });
 
     it('should handle include with threadId parameter', async () => {
